+++ components/inspectit-ocelot-configurationserver/src/test/java/rocks/inspectit/ocelot/rest/file/DirectoryControllerIntTest.java
@@ -6,9 +6,11 @@ import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import rocks.inspectit.ocelot.IntegrationTestBase;
+import rocks.inspectit.ocelot.file.FileData;
 import rocks.inspectit.ocelot.file.FileInfo;
 import rocks.inspectit.ocelot.file.versioning.model.WorkspaceDiff;
 
+import java.io.File;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Optional;
@@ -41,63 +43,6 @@ class DirectoryControllerIntTest extends IntegrationTestBase {
                     .hasSize(1)
                     .extracting(FileInfo::getName, FileInfo::getType, FileInfo::getChildren)
                     .contains(tuple("file.yml", FileInfo.Type.FILE, null));
-        }
-
-
-        @Test
-        public void listLiveWorkspace() {
-            authRest.exchange("/api/v1/files/file.yml", HttpMethod.PUT, null, Void.class);
-
-            ResponseEntity<FileInfo[]> result = authRest.getForEntity("/api/v1/directories/", FileInfo[].class);
-
-            FileInfo[] resultBody = result.getBody();
-            assertThat(resultBody)
-                    .hasSize(1)
-                    .extracting(FileInfo::getName, FileInfo::getType, FileInfo::getChildren)
-                    .contains(tuple("file.yml", FileInfo.Type.FILE, null));
-
-            ResponseEntity<FileInfo[]> resultLive = authRest.getForEntity("/api/v1/directories?version=live", FileInfo[].class);
-
-            FileInfo[] resultBodyLive = resultLive.getBody();
-            assertThat(resultBodyLive)
-                    .hasSize(0);
-        }
-
-        @Test
-        public void listFileFromWorkspace() {
-            authRest.exchange("/api/v1/files/file.yml", HttpMethod.PUT, null, Void.class);
-
-            ResponseEntity<FileInfo[]> result = authRest.getForEntity("/api/v1/directories/", FileInfo[].class);
-
-            FileInfo[] resultBody = result.getBody();
-            assertThat(resultBody)
-                    .hasSize(1)
-                    .extracting(FileInfo::getName, FileInfo::getType, FileInfo::getChildren)
-                    .contains(tuple("file.yml", FileInfo.Type.FILE, null));
-
-            ResponseEntity<Optional> result2 = authRest.getForEntity("/api/v1/files/file.yml", Optional.class);
-
-            Optional<String> resultBody2 = result2.getBody();
-            assertThat(resultBody2.toString()).contains("content=");
-        }
-
-        @Test
-        public void listFileFromLiveWorkspace() {
-
-            authRest.exchange("/api/v1/files/file.yml", HttpMethod.PUT, null, Void.class);
-
-            ResponseEntity<FileInfo[]> result = authRest.getForEntity("/api/v1/directories/", FileInfo[].class);
-
-            FileInfo[] resultBody = result.getBody();
-            assertThat(resultBody)
-                    .hasSize(1)
-                    .extracting(FileInfo::getName, FileInfo::getType, FileInfo::getChildren)
-                    .contains(tuple("file.yml", FileInfo.Type.FILE, null));
-
-            ResponseEntity<Optional> result1 = authRest.getForEntity("/api/v1/files/file.yml?version=live/", Optional.class);
-
-            Optional<String> resultBodyLive = result1.getBody();
-            assertThat(resultBodyLive.toString()).contains("status=INTERNAL_SERVER_ERROR");
         }
     }
